requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
mangum==0.19.0
boto3>=1.35.76
botocore>=1.35.76
